8568: ApplicationFormService
Refactor application form service and also moved some logic to application package service.
