+++ social-middleware/src/application-form/dto/delete-application-form.dto.ts
@@ -1,0 +1,11 @@
+import { IsUUID } from 'class-validator';
+import { ApiProperty } from '@nestjs/swagger';
+
+export class DeleteApplicationFormDto {
+  @ApiProperty({
+    description: 'The applicationId to be deleted',
+    example: '4e4d305b-6de8-4fc8-98cf-fdc4f22111f0',
+  })
+  @IsUUID()
+  applicationId!: string;
+}
