+++ social-middleware/src/application-package/application-package.module.ts
@@ -1,30 +1,48 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
-import { ApplicationPackageController } from './application-package.controller';
-import { ApplicationPackageService } from './application-package.service';
+import { ApplicationFormService } from '../application-form/services/application-form.service';
+import { AccessCodeService } from '../application-form/services/access-code.service';
+import { AuthModule } from 'src/auth/auth.module';
+import { HouseholdModule } from 'src/household/household.module';
+
+import {
+  ApplicationForm,
+  ApplicationFormSchema,
+} from '../application-form/schemas/application-form.schema';
+import {
+  FormParameters,
+  FormParametersSchema,
+} from '../application-form/schemas/form-parameters.schema';
 import {
   ApplicationPackage,
   ApplicationPackageSchema,
-} from './schema/application-package.schema';
-import { ApplicationFormModule } from '../application-form/application-form.module';
-import { HouseholdModule } from '../household/household.module';
-import { AuthModule } from '../auth/auth.module';
-import { SessionUtil } from 'src/common/utils/session.util';
+} from 'src/application-package/schema/application-package.schema';
+import {
+  ScreeningAccessCode,
+  ScreeningAccessCodeSchema,
+} from '../application-form/schemas/screening-access-code.schema';
 
 @Module({
   imports: [
     MongooseModule.forFeature([
-      {
-        name: ApplicationPackage.name,
-        schema: ApplicationPackageSchema,
-      },
+      { name: ApplicationForm.name, schema: ApplicationFormSchema },
+      { name: FormParameters.name, schema: FormParametersSchema },
+      { name: ApplicationPackage.name, schema: ApplicationPackageSchema },
+      { name: ScreeningAccessCode.name, schema: ScreeningAccessCodeSchema },
     ]),
-    ApplicationFormModule,
-    HouseholdModule,
     AuthModule,
+    HouseholdModule,
   ],
-  controllers: [ApplicationPackageController],
-  providers: [ApplicationPackageService, SessionUtil],
-  exports: [ApplicationPackageService],
+  providers: [ApplicationFormService, AccessCodeService],
+  exports: [
+    ApplicationFormService,
+    AccessCodeService,
+    MongooseModule.forFeature([
+      { name: ApplicationForm.name, schema: ApplicationFormSchema },
+      { name: FormParameters.name, schema: FormParametersSchema },
+      { name: ApplicationPackage.name, schema: ApplicationPackageSchema },
+      { name: ScreeningAccessCode.name, schema: ScreeningAccessCodeSchema },
+    ]),
+  ],
 })
 export class ApplicationPackageModule {}
